requirements.txt
Flask==3.0.0
pandas==2.3.3
openpyxl==3.1.5
Pillow==10.1.0
requests==2.31.0

